combine button-section.js and button-widget.js
